pyproject.toml
[project]
name = "check60"
version = "1.3.0"
description = "Utility for fast check your programs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "colorama>=0.4.6",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "pyinstaller>=6.17.0",
    "pyyaml>=6.0.3",
    "undetected-chromedriver>=3.5.5",
]
